+++ packages/plugin-undo-redo/web/src/UndoButton.tsx
@@ -1,28 +1,24 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import type { FunctionComponent } from 'react';
 import React from 'react';
 import classNames from 'classnames';
 import undoIcon from './icons/UndoIcon';
-import { InlineToolbarButton, undo, FORMATTING_BUTTONS } from 'wix-rich-content-editor-common';
+import { InlineToolbarButton, FORMATTING_BUTTONS } from 'wix-rich-content-editor-common';
 import { UNDO_REDO_TYPE } from './types';
-import type {
-  RichContentTheme,
-  Helpers,
-  SetEditorState,
-  GetEditorState,
-} from 'wix-rich-content-common/src';
+import type { RichContentTheme, Helpers } from 'wix-rich-content-common/src';
 import type { TranslationFunction } from 'i18next';
 
 interface UndoButtonProps {
   theme?: RichContentTheme;
   helpers?: Helpers;
-  setEditorState: SetEditorState;
   isMobile?: boolean;
   className?: string;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   config?: any;
   tabIndex?: number;
   t: TranslationFunction;
-  getEditorState: GetEditorState;
+  isDisabled: () => boolean;
+  onClick: (e: any & { ref?: any; render?: any }) => void;
 }
 
 const UndoButton: FunctionComponent<UndoButtonProps> = props => {
@@ -35,18 +31,12 @@ const UndoButton: FunctionComponent<UndoButtonProps> = props => {
     config,
     tabIndex,
     t,
-    getEditorState,
-    setEditorState,
+    isDisabled,
+    onClick,
   } = props;
-  const editorState = getEditorState();
   const combinedClassName = classNames(theme.undo, className);
   const icon = config?.toolbar?.icons?.Undo || undoIcon;
-  const disabled = editorState?.getUndoStack()?.isEmpty?.() || !editorState;
-
-  const onClick = event => {
-    event.stopPropagation();
-    setEditorState(undo(getEditorState()));
-  };
+  const disabled = isDisabled();
 
   if (isMobile)
     return (