+++ packages/plugin-undo-redo/web/src/createToolbar.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import React from 'react';
 import { BUTTON_TYPES, FORMATTING_BUTTONS, undo, redo } from 'wix-rich-content-editor-common';
 import UndoIcon from './icons/UndoIcon';
@@ -10,74 +11,78 @@ import type {
   TranslationFunction,
   GetEditorState,
   SetEditorState,
-  Pubsub,
+  Helpers,
 } from 'wix-rich-content-common';
 import type { UndoRedoPluginEditorConfig } from './types';
+import { UNDO_REDO_TYPE } from './types';
 
 const createToolbar: CreatePluginToolbar = ({
   t,
   getEditorState,
   setEditorState,
   settings,
+  helpers,
 }: {
   t: TranslationFunction;
   getEditorState: GetEditorState;
   setEditorState: SetEditorState;
   settings: UndoRedoPluginEditorConfig;
-  commonPubsub: Pubsub;
+  helpers: Helpers;
 }) => {
+  const onUndoClick = (e: any & { ref?: any; render?: any }) => {
+    e.preventDefault();
+    setEditorState(undo(getEditorState()));
+    helpers?.onPluginAddSuccess?.(UNDO_REDO_TYPE, 'UNDO', {});
+  };
+
+  const onRedoClick = (e: any & { ref?: any; render?: any }) => {
+    e.preventDefault();
+    setEditorState(redo(getEditorState()));
+    helpers?.onPluginAddSuccess?.(UNDO_REDO_TYPE, 'REDO', {});
+  };
+
+  const isUndoDisabled = () => !!getEditorState()?.getUndoStack().isEmpty();
+
+  const isRedoDisabled = () => !!getEditorState()?.getRedoStack().isEmpty();
+
   return {
     TextButtonMapper: () => ({
       [FORMATTING_BUTTONS.UNDO]: {
         component: props => (
-          <UndoButton
-            t={t}
-            getEditorState={getEditorState}
-            setEditorState={setEditorState}
-            {...props}
-          />
+          <UndoButton t={t} onClick={onUndoClick} isDisabled={isUndoDisabled} {...props} />
         ),
         externalizedButtonProps: {
           type: BUTTON_TYPES.BUTTON,
           getLabel: () => '',
           isActive: () => false,
-          isDisabled: () => getEditorState().getUndoStack().isEmpty(),
+          isDisabled: isUndoDisabled,
           tooltip: t('UndoButton_Tooltip'),
           getIcon: () => settings?.toolbars?.icons?.Undo || UndoIcon,
-          onClick: e => {
-            e.preventDefault();
-            setEditorState(undo(getEditorState()));
-          },
+          onClick: onUndoClick,
         },
       },
       [FORMATTING_BUTTONS.REDO]: {
         component: props => (
-          <RedoButton
-            t={t}
-            getEditorState={getEditorState}
-            setEditorState={setEditorState}
-            {...props}
-          />
+          <RedoButton t={t} onClick={onRedoClick} isDisabled={isRedoDisabled} {...props} />
         ),
         externalizedButtonProps: {
           getLabel: () => '',
           type: BUTTON_TYPES.BUTTON,
           isActive: () => false,
-          isDisabled: () => getEditorState().getRedoStack().isEmpty(),
+          isDisabled: isRedoDisabled,
           tooltip: t('RedoButton_Tooltip'),
           getIcon: () => settings?.toolbars?.icons?.Redo || RedoIcon,
-          onClick: e => {
-            e.preventDefault();
-            setEditorState(redo(getEditorState()));
-          },
+          onClick: onRedoClick,
         },
       },
     }),
     InsertButtons: createInsertButtons({
       t,
-      getEditorState,
-      setEditorState,
       settings,
+      onUndoClick,
+      onRedoClick,
+      isUndoDisabled,
+      isRedoDisabled,
     }),
     name: 'undo-redo',
   };