+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 
 - `media-plugins`
   - [#3474](https://github.com/wix/ricos/pull/3474) raise generic error toast when local load fails and raise error in console
+- `ricos-content`
+  - [#3499](https://github.com/wix/ricos/pull/3499) fix `fromDraft` `inlineStyles` to `textDecoration` converter
 
 ### :house: Internal